Moved MethodBodyTypechecker into its own file, implemented more semantic error reporting.
